Move tree item, directory, entry and iter to separate files
The only reason to have them in the same file was because each item
kept a reference to the tree. Since we don't use it, and most likely
never will, there's no real cost to separating it out like this, but
it does help to simplify cmenu-tree.c which is a huge, tangled mess.
This also helps to improve encapsulation overall, which is severely
lacking in cmenu.