Ignore layout and default layout
These are optional parts of the spec that we ultimately ignore anyway. Quite a bit of code (~500 lines) is dedicated just to handling this.

This is now replaced with a much simpler code that merely flattens the categories and sorts them. We currently do these in the cinnamon menu applet and the cinnamon-menu-editor as well, but this will allow us to simplify that code as well. These changes should be backwards-compatible with current implementations, but they will have some redundant code that we'll want to remove for the sake of simplicity and efficiency.