gmenu-tree.c: Return a basename if no appinfo is available when sorting.
This is a workaround: if a desktop file is edited and becomes
invalid, the tree rebuilds and re-loading that file initially fails
and gets added to our "retry later" list (see b5495c4176e). The
problem is, mimeinfo never updates in this case, and we end up
with basically an invalid desktop entry with no appinfo.

This is mainly to silence a lot of runtime warnings because the
sorting fails.