Improve integration of flatpak apps
This adds the suffix ":flatpak" to the desktop id, so that Cinnamon can keep Flatpak apps separate from system repositories' apps.

It's done by adding a wrapper for GDesktopAppInfo that only changes the desktop file id and the StartupWMClass
while making sure it's used where it's supposed to be used.