packaging: Forbid compilation if the symbols file is outdated
An outdated symbols file results in automatic versioned dependencies
in projects which use this library.

When these projects are built, shlibs:Depends cannot find a version
number for the set of symbols so it uses the full version of the
library package, including suffixes such as +virginia.

We end up with depends lib (>== 6.0.0+virginia) instead of
lib (>== 6.0.0).